+++ main.go
@@ -6,6 +6,7 @@ import (
 	"sort"
 	"strings"
 	"syscall"
+	"time"
 
 	clipboard "github.com/atotto/clipboard"
 	figure "github.com/common-nighthawk/go-figure"
@@ -43,21 +44,27 @@ func main() {
 	}
 
 	if isList(command) {
-		DoList(cli)
+		DoList(cli, "")
+	} else if isSearch(command) {
+		searchTerm := cli.GetStringOrDefault(command, "")
+		DoList(cli, searchTerm)
 	} else if isVersion(command) {
 		DoVersion(cli)
-	} else if isDescribe(command) {
-		DoDescribe(cli)
 	} else if isPut(command, cli) {
 		DoPut(cli)
 	} else if isEncrypt(command) {
 		DoEncrypt(cli)
 	} else if isDecrypt(command) {
 		DoDecrypt(cli)
+	} else if isDescribe(command) {
+		DoDescribe(cli)
 	} else if isUpdate(command) {
-		DoUpdateDescription(cli)
+		DoUpdate(cli)
+	} else if isRename(command) {
+		DoRename(cli)
 	} else if isGet(command, cli) {
 		DoGet(cli)
+		DoDescribe(cli)
 	} else if isDelete(command) {
 		DoDelete(cli)
 	} else if command != "" {
@@ -94,12 +101,20 @@ func isUpdate(command string) bool {
 	return command == "update"
 }
 
+func isRename(command string) bool {
+	return command == "rename"
+}
+
 func isVersion(command string) bool {
 	return command == "version"
 }
 
 func isList(command string) bool {
 	return command == "ls"
+}
+
+func isSearch(command string) bool {
+	return command == "search"
 }
 
 func isDescribe(command string) bool {
@@ -157,13 +172,6 @@ func DoVerify(c *cli.CLI, printFailuresToStdOut bool) bool {
 	messages = append(messages, fmt.Sprintf("%v   : %v, exists=%v\n", KP_FILE, kpFilename, filenameExists))
 	messages = append(messages, fmt.Sprintf("%v    : %v, exists=%v\n", KP_KEY, privateKeyFilename, privateKeyExists))
 
-	// if !filenameExists {
-	// 	// line := fmt.Sprintf("KP_FILE '%v' does not exist.\n", kpFilename)
-	// 	// messages = append(messages, line)
-	// 	overallValid = false
-	// } else {
-	// 	// fmt.Printf("KP_FILE '%v' exists.\n", kpFilename)
-	// }
 	if !privateKeyExists {
 		overallValid = false
 		messages = append(messages, "\nEncryption key does not exist, try the following\n\n")
@@ -216,30 +224,13 @@ func DoInfo(c *cli.CLI) {
 	msg = strings.ReplaceAll(msg, "TOKEN_DEFAULT_SSH_COMMAND", GetSSHCommand(privKey))
 	fmt.Printf("\n%v\n", msg)
 
-	// t := NewTerminal()
-
-	// sysInfo := goutils.NewSysInfo()
-
-	// fmt.Printf("RAM         : %v\n", sysInfo.RAM)
-	// fmt.Printf("CPU         : %v\n", sysInfo.CPU)
-	// fmt.Printf("Cores	    : %v\n", runtime.NumCPU())
-	// fmt.Printf("Disk        : %v\n", sysInfo.Disk)
-	// fmt.Printf("Hostname    : %v\n", sysInfo.Hostname)
-	// fmt.Printf("GOOS        : %v\n", runtime.GOOS)
-	// fmt.Printf("GOARCH      : %v\n", runtime.GOARCH)
-	// fmt.Printf("GOMAXPROC   : %v\n", runtime.GOMAXPROCS)
-	// fmt.Printf("Columns     : %v\n", t.Width())
-	// fmt.Printf("IsMacOS     : %v\n", sysInfo.IsMacOS())
-	// fmt.Printf("IsLinux     : %v\n", sysInfo.IsLinux())
-	// fmt.Printf("IsWindows   : %v\n", sysInfo.IsWindows())
-
 }
 
 func DoGet(c *cli.CLI) {
 	command := c.GetCommand()
 	key := c.GetStringOrDie(command)
 	db := LoadDB()
-	entry, exists := db.Get(key)
+	entry, exists := db.GetDecrypted(key)
 	if exists {
 		value := entry.Value
 		err := clipboard.WriteAll(value)
@@ -259,21 +250,20 @@ func DoPut(c *cli.CLI) {
 	db := LoadDB()
 	command := c.GetCommand()
 	key := c.GetStringOrDie(command)
-	if len(key) > 25 {
+	if len(key) > 125 {
 		fmt.Printf("Error, key must be <= 25 characters.\n")
 		os.Exit(1)
 	} else if len(key) == 0 {
 		fmt.Printf("Error, key cannot be empty.")
 		os.Exit(1)
 	}
-	entry, exists := db.Get(key)
-	var description string
-	if exists {
-		description = entry.Description
-	} else {
-		description = ""
-	}
-	description = c.GetStringOrDefault("-d", description)
+	entry, _ := db.GetDecrypted(key)
+	entry.Description = c.GetStringOrDefault("-description", entry.Description)
+	entry.Type = c.GetStringOrDefault("-type", entry.Type)
+	entry.Notes = c.GetStringOrDefault("-note", entry.Notes)
+	entry.Url = c.GetStringOrDefault("-url", entry.Url)
+	entry.Username = c.GetStringOrDefault("-username", entry.Username)
+
 	password := ""
 	if c.IndexOf("-value") > -1 {
 		password = c.GetStringOrDefault("-value", "")
@@ -285,17 +275,44 @@ func DoPut(c *cli.CLI) {
 		bytePassword, _ := terminal.ReadPassword(int(syscall.Stdin))
 		password = string(bytePassword)
 	}
-	value := password
-	db.Put(key, value, description)
+	entry.Value = password
+	db.Put(entry)
 	db.Save()
 }
 
-func DoUpdateDescription(c *cli.CLI) {
+func DoUpdate(c *cli.CLI) {
 	db := LoadDB()
 	command := c.GetCommand()
 	key := c.GetStringOrDie(command)
-	description := c.GetStringOrDie(key)
-	db.UpdateDescription(key, description)
+	entry, _ := db.GetDecrypted(key)
+	entry.Description = c.GetStringOrDefault("-description", entry.Description)
+	entry.Type = c.GetStringOrDefault("-type", entry.Type)
+	entry.Notes = c.GetStringOrDefault("-note", entry.Notes)
+	entry.Url = c.GetStringOrDefault("-url", entry.Url)
+	entry.Username = c.GetStringOrDefault("-username", entry.Username)
+	db.Put(entry)
+	db.Save()
+}
+
+func DoRename(c *cli.CLI) {
+	db := LoadDB()
+	command := c.GetCommand()
+	old_key := c.GetStringOrDie(command)
+	entry, exists := db.GetDecrypted(old_key)
+	if !exists {
+		fmt.Printf("No such entry '%v'\n", old_key)
+		os.Exit(1)
+	}
+
+	new_key := c.GetStringOrDie(old_key)
+	_, other_exists := db.GetDecrypted(new_key)
+	if other_exists {
+		fmt.Printf("Entry '%v' already exists.\n", new_key)
+		os.Exit(1)
+	}
+	entry.Key = new_key
+	db.Delete(old_key)
+	db.Put(entry)
 	db.Save()
 }
 
@@ -303,13 +320,23 @@ func DoDescribe(c *cli.CLI) {
 	db := LoadDB()
 	command := c.GetCommand()
 	key := c.GetStringOrDie(command)
-	description := c.GetStringOrDie(key)
-	value, _ := db.Get(key)
-	db.Put(key, value.Value, description)
-	db.Save()
+	entry, exists := db.GetDecrypted(key)
+	if exists {
+		fmt.Printf("Key          : %v\n", entry.Key)
+		fmt.Printf("Description  : %v\n", entry.Description)
+		fmt.Printf("Username     : %v\n", entry.Username)
+		fmt.Printf("Url          : %v\n", entry.Url)
+		fmt.Printf("Created      : %v\n", entry.Created.Format(time.RFC822Z))
+		fmt.Printf("Last Updated : %v\n", entry.LastUpdated.Format(time.RFC822Z))
+		fmt.Printf("Type         : %v\n", entry.Type)
+		fmt.Printf("Notes        : %v\n", entry.Notes)
+	} else {
+		fmt.Printf("Error, no entry '%v'\n", entry.Key)
+	}
+
 }
 
-func DoList(c *cli.CLI) {
+func DoList(c *cli.CLI, searchTerm string) {
 	db := LoadDB()
 	data := db.GetData()
 	if len(data.Entries) == 0 {
@@ -321,62 +348,81 @@ func DoList(c *cli.CLI) {
 		}
 
 		keys := make([]string, 0)
-		for key := range data.Entries {
+		max_key := len("Key") + 1
+		max_url := len("Url") + 1
+		max_username := len("Username") + 1
+		max_type := len("Type") + 1
+		max_description := len("Description") + 1
+		max_notes := len("Notes") + 1
+
+		for key, entry := range data.Entries {
 			keys = append(keys, key)
+			max_key = goutils.Max(len(entry.Key)+1, max_key)
+			max_url = goutils.Max(len(entry.Url)+1, max_url)
+			max_username = goutils.Max(len(entry.Username)+1, max_username)
+			max_type = goutils.Max(len(entry.Type)+1, max_type)
+			max_description = goutils.Max(len(entry.Description)+1, max_description)
+			// max_notes = goutils.Max(len(entry.Notes)+1, max_notes)
 		}
 		sort.Strings(keys)
 
-		key_width := 50
-		desc_width := 50
-		date_width := 25
-		width := key_width + desc_width + (2 * date_width) + 4
+		date_width := len(time.Now().Format(time.RFC822Z)) + 1
+		width := max_key + max_url + max_username + max_type + max_description + max_notes + (2 * date_width) + 8
 
-		line := strings.Repeat("-", width) + "\n"
-		fmt.Print(line)
-		header := fmt.Sprintf("| Key%v| Description%v| Updated%v| Created%v|\n", strings.Repeat(" ", key_width-len("Key")-1), strings.Repeat(" ", desc_width-len("Description")-1), strings.Repeat(" ", date_width-len("Updated")-1), strings.Repeat(" ", date_width-len("Created")-1))
-		fmt.Print(header)
-		fmt.Print(line)
-		max_key_length := key_width - 5
-		max_description_length := desc_width - 5
-		for _, key := range keys {
-			entry := data.Entries[key]
-			created := entry.Created.Format("January 2, 2006")
-			updated := entry.LastUpdated.Format("January 2, 2006")
+		line := strings.Repeat("-", width+1)
+		fmt.Println(line)
+
+		entry_line := fmt.Sprintf("|%v|%v|%v|%v|%v|%v|%v|%v|",
+			goutils.RPadToFixedLength("Key", " ", max_key),
+			goutils.RPadToFixedLength("Username", " ", max_username),
+			goutils.RPadToFixedLength("Url", " ", max_url),
+			goutils.RPadToFixedLength("Type", " ", max_type),
+			goutils.RPadToFixedLength("Description", " ", max_description),
+			goutils.RPadToFixedLength("Notes", " ", max_notes),
+			goutils.RPadToFixedLength("Updated", " ", date_width),
+			goutils.RPadToFixedLength("Created", " ", date_width))
+		fmt.Println(entry_line)
+		fmt.Println(line)
+
+		for _, entry := range db.GetEntriesSortedByUpdatedThenKey() {
+
+			if searchTerm != "" {
+				found := strings.Contains(entry.Key, searchTerm)
+				found = found || strings.Contains(entry.Description, searchTerm)
+				found = found || strings.Contains(entry.Notes, searchTerm)
+				found = found || strings.Contains(entry.Type, searchTerm)
+				found = found || strings.Contains(entry.Username, searchTerm)
+				found = found || strings.Contains(entry.Url, searchTerm)
+				if !found {
+					continue
+				}
+
+			}
+			notes := entry.Notes
+			if notes != "" {
+				notes = "***"
+			}
+			created := entry.Created.Format(time.RFC822)
+			updated := entry.LastUpdated.Format(time.RFC822)
 			desc := entry.Description
 
-			if len(key) > max_key_length {
-				key = key[0:max_key_length] + "..."
+			if len(desc) > max_description {
+				desc = desc[0:max_description-3] + "..."
 			}
 
-			if desc == "" {
-				desc = "No description."
-			}
-			if len(desc) > max_description_length {
-				desc = desc[0:max_description_length] + "..."
-			}
+			entry_line := fmt.Sprintf("|%v|%v|%v|%v|%v|%v|%v|%v",
+				goutils.RPadToFixedLength(entry.Key, " ", max_key),
+				goutils.RPadToFixedLength(entry.Username, " ", max_username),
+				goutils.RPadToFixedLength(entry.Url, " ", max_url),
+				goutils.RPadToFixedLength(entry.Type, " ", max_type),
+				goutils.RPadToFixedLength(desc, " ", max_description),
+				goutils.RPadToFixedLength(notes, " ", max_notes),
+				goutils.RPadToFixedLength(created, " ", date_width),
+				goutils.RPadToFixedLength(updated, " ", date_width))
+			fmt.Println(entry_line)
 
-			keyExtra := key_width - 1 - len(key)
-			descExtra := desc_width - 1 - len(desc)
-			updatedExtra := date_width - 1 - len(updated)
-			createdExtra := date_width - 1 - len(created)
-
-			if keyExtra < 0 {
-				keyExtra = 0
-			}
-			if descExtra < 0 {
-				descExtra = 0
-			}
-			if updatedExtra < 0 {
-				updatedExtra = 0
-			}
-			if createdExtra < 0 {
-				createdExtra = 0
-			}
-
-			entry_line := fmt.Sprintf("| %v%v| %v%v| %v%v| %v%v|\n", key, strings.Repeat(" ", keyExtra), desc, strings.Repeat(" ", descExtra), updated, strings.Repeat(" ", updatedExtra), created, strings.Repeat(" ", createdExtra))
-			fmt.Print(entry_line)
 		}
-		fmt.Print(line)
+		fmt.Println(line)
 	}
 }
 
@@ -388,7 +434,7 @@ func DoDelete(c *cli.CLI) {
 		fmt.Printf("%v\n", USAGE)
 	}
 	db := LoadDB()
-	_, exists := db.Get(key)
+	_, exists := db.GetDecrypted(key)
 	if !exists {
 		fmt.Printf("Error, '%v' does not exist.\n", key)
 		os.Exit(1)