+++ constants.go
@@ -3,7 +3,7 @@ package main
 import "strings"
 
 // VERSION is the number of this beast
-const VERSION = "0.0.8"
+const VERSION = "0.0.9"
 const DB_VERSION = "0.0.1"
 
 // KP_FILE the key for the env var pointint to the file we load/save
@@ -25,13 +25,23 @@ Usage:
 The commands are:
 
     ls                                          list keys
-    put <key> (-value VALUE) (-d description)   save "key/value" (read stdin if "-value" is unspecified)
+
+    put <key> (-value VALUE)                    save "key/value" (read stdin if "-value" is unspecified)
     get <key> (-stdout)                         retrieve key/value to clipboard (or -stdout)
-    update <key> "description"                  update the descrption of the key/pair
+
+    update <key>                                update metadata on the key
+         -description                   
+         -type
+         -url
+         -username
+         -note
+
+    search <query>                              return entries that match the search
+
     rm <key>                                    permanently remove "key"
 
-    encrypt <value>                             encrypt the value using the current key
-    decrypt <value>                             decrypt the value using the current key
+    encrypt <value>                             encrypt the value using the current openssh key
+    decrypt <value>                             decrypt the value using the current openssh key
 
     info                                        review environment variables used
     verify                                      check encryption keys exist and work