+++ main_test.go
@@ -8,16 +8,16 @@ import (
 	cli "github.com/simonski/cli"
 )
 
-func main_test(t *testing.T) {
+func TestMain(t *testing.T) {
 
 	command := "fooo"
 	cli := cli.New(os.Args)
 	if command == "test" {
 		db := LoadDB()
 		value := cli.GetStringOrDie(command)
-		valueEnc := db.Encrypt(value)
+		valueEnc, _ := db.Encrypt(value)
 		fmt.Printf("Encrypt('%v')=\n%v\n", value, valueEnc)
-		valueDec := db.Decrypt(valueEnc)
+		valueDec, _ := db.Decrypt(valueEnc)
 		fmt.Printf("\n\nDecrypt\n '%v'\n", valueDec)
 		os.Exit(0)
 	}