+++ objects.go
@@ -3,8 +3,9 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
+	"sort"
 	"time"
 
 	goutils "github.com/simonski/goutils"
@@ -29,6 +30,10 @@ type DBEntry struct {
 	Key         string    `json:"key"`
 	Value       string    `json:"value"`
 	Description string    `json:"description"`
+	Notes       string    `json:"notes"`
+	Username    string    `json:"username"`
+	Url         string    `json:"url"`
+	Type        string    `json:"type"`
 	LastUpdated time.Time `json:"lastUpdated"`
 	Created     time.Time `json:"created"`
 }
@@ -58,7 +63,7 @@ func (cdb *KPDB) Load(filename string, privKey string) bool {
 			cdb.data = db
 		} else {
 			db := DB{}
-			bytes, _ := ioutil.ReadAll(jsonFile)
+			bytes, _ := io.ReadAll(jsonFile)
 
 			// let's see what sort of DB this is
 			// pre-version was a map of Entries
@@ -68,6 +73,7 @@ func (cdb *KPDB) Load(filename string, privKey string) bool {
 			if db.Version == "" {
 				var data map[string]DBEntry
 				json.Unmarshal(bytes, &data)
+
 				db.Entries = data
 				db.Version = DB_VERSION
 				cdb.data = db
@@ -79,10 +85,35 @@ func (cdb *KPDB) Load(filename string, privKey string) bool {
 				// upgrade()
 				cdb.data = db
 			}
+
+			for k, v := range cdb.data.Entries {
+				if v.Key == "" {
+					v.Key = k
+					cdb.data.Entries[k] = v
+				}
+			}
+
 		}
 	}
 
 	return true
+}
+
+func (cdb *KPDB) GetEntriesSortedByUpdatedThenKey() []DBEntry {
+
+	entries := make([]DBEntry, 0)
+	for _, e := range cdb.data.Entries {
+		entries = append(entries, e)
+	}
+
+	sort.Slice(entries, func(a int, b int) bool {
+		entryA := entries[a]
+		entryB := entries[b]
+		return entryA.LastUpdated.After(entryB.LastUpdated) && entryA.Key < entryB.Key
+	})
+
+	return entries
+
 }
 
 // Clear empties the db (without saving it)
@@ -94,7 +125,7 @@ func (cdb *KPDB) Clear() {
 func (cdb *KPDB) Save() bool {
 	data := cdb.data
 	file, _ := json.MarshalIndent(data, "", " ")
-	err := ioutil.WriteFile(cdb.Filename, file, 0644)
+	err := os.WriteFile(cdb.Filename, file, 0644)
 	if err != nil {
 		fmt.Printf("%v", err)
 	}
@@ -107,11 +138,10 @@ func (cdb *KPDB) GetData() DB {
 }
 
 // Get returns the (DBEntry, bool) indicating it exists (or not)
-func (cdb *KPDB) Get(key string) (DBEntry, bool) {
+func (cdb *KPDB) GetDecrypted(key string) (DBEntry, bool) {
 	entry, exists := cdb.data.Entries[key]
 	if exists {
-		decValue := entry.Value
-		decValue = cdb.Decrypt(entry.Value)
+		decValue, _ := cdb.Decrypt(entry.Value)
 		entry.Value = decValue
 	}
 	return entry, exists
@@ -119,35 +149,24 @@ func (cdb *KPDB) Get(key string) (DBEntry, bool) {
 
 // Get returns the (DBEntry, bool) indicating it exists (or not)
 func (cdb *KPDB) UpdateDescription(key string, description string) (DBEntry, bool) {
-	entry, exists := cdb.Get(key)
-	cdb.Put(entry.Key, entry.Value, description)
-	if exists {
-		entry.Description = description
-	}
+	entry, exists := cdb.GetDecrypted(key)
+	entry.Description = description
+	cdb.Put(entry)
 	return entry, exists
 }
 
 // Put stores (or replaces) the key/value pair
-func (cdb *KPDB) Put(key string, value string, description string) {
-	entry, exists := cdb.data.Entries[key]
-	encValue := value
-	encValue = cdb.Encrypt(value)
-	if exists {
-		if value != "" {
-			entry.Value = encValue
-		}
-		entry.LastUpdated = time.Now()
-		if description != "" {
-			entry.Description = description
-		}
-		cdb.data.Entries[key] = entry
+func (cdb *KPDB) Put(entry_in DBEntry) {
+	entry, exists := cdb.data.Entries[entry_in.Key]
+	encValue, _ := cdb.Encrypt(entry_in.Value)
+	entry_in.Value = encValue
+	if !exists {
+		entry_in.Created = entry.Created
 	} else {
-		entry = DBEntry{Key: key, Value: encValue, Created: time.Now(), LastUpdated: time.Now()}
-		if description != "" {
-			entry.Description = description
-		}
-		cdb.data.Entries[key] = entry
+		entry_in.Created = time.Now()
 	}
+	entry_in.LastUpdated = time.Now()
+	cdb.data.Entries[entry_in.Key] = entry_in
 }
 
 // Delete removes the key/value pair from the DB
@@ -156,13 +175,11 @@ func (cdb *KPDB) Delete(key string) {
 }
 
 // Encrypt helper function encrypts with public key
-func (cdb *KPDB) Encrypt(value string) string {
-	encrypted := crypto.EncryptWithPrivateKeyFilename(value, cdb.PrivateKeyFilename)
-	return encrypted
+func (cdb *KPDB) Encrypt(value string) (string, error) {
+	return crypto.EncryptWithPrivateKeyFilename(value, cdb.PrivateKeyFilename)
 }
 
 // Decrypt helper function decrypts with private key
-func (cdb *KPDB) Decrypt(value string) string {
-	decrypted := crypto.DecryptWithPrivateKeyFilename(value, cdb.PrivateKeyFilename)
-	return decrypted
+func (cdb *KPDB) Decrypt(value string) (string, error) {
+	return crypto.DecryptWithPrivateKeyFilename(value, cdb.PrivateKeyFilename)
 }