+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2022-08-03
+
+0.0.9
+go 1.19
+staticcheck
+UX improvements for put, get and list
+added "notes, username, url" to the model
+
 2022-04-02
 0.0.8
 Added version to .json schema to allow for future upgrade logic.